LayoutTests/platform/glib/imported/w3c/web-platform-tests/css/css-cascade/layer-rules-cssom-expected.txt


PASS Basic layer block name
PASS Anonymous layer block name
PASS Basic layer statement name
PASS Layer statement with multiple names
PASS Nested layer block names
PASS Nested layer statement name lists
FAIL Import into anonymous layer assert_equals: expected (string) "" but got (undefined) undefined
FAIL Import into named layer assert_equals: expected (string) "foo" but got (undefined) undefined
FAIL Import without layer assert_equals: expected (object) null but got (undefined) undefined
